refactor: move validateMcpTemplate to core/mcp.ts
